+++ src/main/java/pl/coderslab/projectwybankuj/repository/MortgageRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import pl.coderslab.projectwybankuj.entity.Mortgage;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,14 +15,14 @@ public interface MortgageRepository extends CrudRepository<Mortgage, Long> {
     List<Mortgage> findAllByBankId(Long bankId);
 
     @Query("SELECT m FROM Mortgage m WHERE m.minCreditAmount <= :amount AND m.maxCreditAmount >= :amount AND m.maxCreditPeriod >= :creditPeriod AND m.minBorrowerAge <= :age AND m.maxBorrowerAge >= :age AND m.contributionPercent <= :contributionPercent AND m.maxBorrowerAge - :age >= :creditPeriod / 12")
-    List<Mortgage> findAllByParameters(int amount, int creditPeriod, int age, double contributionPercent);
+    List<Mortgage> findAllByParameters(int amount, int creditPeriod, int age, BigDecimal contributionPercent);
 
     @Query("SELECT m FROM Mortgage m WHERE m.minCreditAmount <= :amount AND m.maxCreditAmount >= :amount AND m.maxCreditPeriod >= :creditPeriod AND m.minBorrowerAge <= :age AND m.maxBorrowerAge >= :age AND m.insurance = 0.0 AND m.contributionPercent <= :contributionPercent AND m.maxBorrowerAge - :age >= :creditPeriod / 12")
-    List<Mortgage> findAllByParametersWithoutInsurance(int amount, int creditPeriod, int age, double contributionPercent);
+    List<Mortgage> findAllByParametersWithoutInsurance(int amount, int creditPeriod, int age, BigDecimal contributionPercent);
 
     @Query("SELECT m FROM Mortgage m WHERE m.minCreditAmount <= :amount AND m.maxCreditAmount >= :amount AND m.maxCreditPeriod >= :creditPeriod AND m.minBorrowerAge <= :age AND m.maxBorrowerAge >= :age AND m.serviceCharge = 0.0 AND m.contributionPercent <= :contributionPercent AND m.maxBorrowerAge - :age >= :creditPeriod / 12")
-    List<Mortgage> findAllByParametersWithoutServiceCharge(int amount, int creditPeriod, int age, double contributionPercent);
+    List<Mortgage> findAllByParametersWithoutServiceCharge(int amount, int creditPeriod, int age, BigDecimal contributionPercent);
 
     @Query("SELECT m FROM Mortgage m WHERE m.minCreditAmount <= :amount AND m.maxCreditAmount >= :amount AND m.maxCreditPeriod >= :creditPeriod AND m.minBorrowerAge <= :age AND m.maxBorrowerAge >= :age AND m.insurance = 0.0 AND m.serviceCharge = 0.0 AND m.contributionPercent <= :contributionPercent AND m.maxBorrowerAge - :age >= :creditPeriod / 12")
-    List<Mortgage> findAllByParametersWithoutInsuranceAndServiceCharge(int amount, int creditPeriod, int age, double contributionPercent);
+    List<Mortgage> findAllByParametersWithoutInsuranceAndServiceCharge(int amount, int creditPeriod, int age, BigDecimal contributionPercent);
 }
